+++ src/components/custom/index.js
@@ -1,0 +1,5 @@
+import Details from "./Details"
+
+export default {
+    Details
+}