+++ src/components/node/Node.js
@@ -4,64 +4,92 @@ import {Handle} from 'react-flow-renderer';
 
 import {KEY_VALUES, NODE} from "../../lib/configParsing";
 import {getShape} from "./nodeUtil";
-
+import customComponents from '../custom';
 
 //import cylinder from "../../assets/cylinder.svg"
 
-export default memo(({data, isConnectable}) => {
+function shape({ shape, fill, stroke, strokeWidth, strokeDashArray, topText, fontSize, width, height, image, title, label }) {
+    let element = getShape(shape, fill, stroke, strokeWidth, strokeDashArray);
+    if (!element) {
+        return null;
+    }
+    return <>
+        {topText &&
+            <p style={{ margin: 0, padding: 0, textAlign: "center", fontSize: fontSize - 1 }}>{topText}</p>
+        }
+        <svg style={{ width: width, height: height }}>
+            {element}
 
-    let width = data[KEY_VALUES[NODE].WIDTH.id];
-    let height = data[KEY_VALUES[NODE].HEIGHT.id];
+            <svg /*viewBox={`0 ${s} ${width} ${height}`}*/ width={width} height={height}>
+                {image &&
+                    (getShape(image) || <image key={Math.random()} href={image} width={width} height={height} />)
+                }
+            </svg>
 
-    let fontsize = data[KEY_VALUES[NODE].FONTSIZE.id];
+            <text fontSize={fontSize}>
+                {title &&
+                    <tspan key={Math.random()} x="50%"
+                        y={(strokeWidth || 1) + fontSize}
+                        dominantBaseline="middle" textAnchor="middle">{title}</tspan>
+                }
+                {label &&
+                    label.split("\n").map((e, i) => {
+                        if (i !== 0) {
+                            return <tspan key={i} x="50%" dy={fontSize} dominantBaseline="middle"
+                                textAnchor="middle">{e}</tspan>
+                        } else {
+                            return <tspan key={i} x="50%"
+                                y={50 - ((label.split("\n").length - 1) * height / fontSize / 2) + "%"}
+                                dominantBaseline="middle" textAnchor="middle">{e}</tspan>
+                        }
+                    })}
+            </text>
+        </svg>
+    </>
+}
 
-    let element = getShape(
-        data[KEY_VALUES[NODE].SHAPE.id], data[KEY_VALUES[NODE].FILL.id], data[KEY_VALUES[NODE].STROKE.id], data[KEY_VALUES[NODE].STROKE_WIDTH.id], data[KEY_VALUES[NODE].STROKE_DASHARRAY.id]
-    );
+function component(props){
+    const {shape} = props;
+    let Element = customComponents[shape];
+    if (Element) {
+        return <Element {...props}></Element>
+    }
+}
 
+export default memo(({ data, isConnectable }) => {
+
+    let myShape = shape({
+        shape: data[KEY_VALUES[NODE].SHAPE.id],
+        fill: data[KEY_VALUES[NODE].FILL.id],
+        stroke: data[KEY_VALUES[NODE].STROKE.id],
+        strokeWidth: data[KEY_VALUES[NODE].STROKE_WIDTH.id],
+        strokeDashArray: data[KEY_VALUES[NODE].STROKE_DASHARRAY.id],
+        topText: data[KEY_VALUES[NODE].TOP_TEXT.id],
+        fontSize: data[KEY_VALUES[NODE].FONTSIZE.id],
+        width: data[KEY_VALUES[NODE].WIDTH.id],
+        height: data[KEY_VALUES[NODE].HEIGHT.id],
+        image: data[KEY_VALUES[NODE].IMAGE.id],
+        title: data[KEY_VALUES[NODE].TITLE.id],
+        label: data[KEY_VALUES[NODE].LABEL.id]
+    })
+
+    let myCustomComponent = component({
+        shape: data[KEY_VALUES[NODE].SHAPE.id],
+        topText: data[KEY_VALUES[NODE].TOP_TEXT.id],
+        fontSize: data[KEY_VALUES[NODE].FONTSIZE.id],
+        width: data[KEY_VALUES[NODE].WIDTH.id],
+        height: data[KEY_VALUES[NODE].HEIGHT.id],
+        image: data[KEY_VALUES[NODE].IMAGE.id],
+        title: data[KEY_VALUES[NODE].TITLE.id],
+        label: data[KEY_VALUES[NODE].LABEL.id]
+    })
+
+    // TODO refactor this: should be either a shape OR a custom component
     return (
         <>
-            {data[KEY_VALUES[NODE].TOP_TEXT.id] &&
-                <p style={{margin: 0, padding: 0, textAlign: "center", fontSize: fontsize-1}}>{data[KEY_VALUES[NODE].TOP_TEXT.id]}</p>
-            }
-            <svg style={{width: width, height: height}}>
+            {myShape}
 
-                <svg style={{width: width}} key={Math.random()}>
-                    {
-                        element
-                    }
-                </svg>
-
-                <svg /*viewBox={`0 ${s} ${width} ${height}`}*/ width={width} height={height}>
-                    {data[KEY_VALUES[NODE].IMAGE.id] &&
-                        (getShape(data[KEY_VALUES[NODE].IMAGE.id]) ||
-                            <image key={Math.random()} href={data[KEY_VALUES[NODE].IMAGE.id]}
-                                   width={width} height={height}/>
-                        )
-                    }
-                </svg>
-
-
-                <text fontSize={fontsize}>
-                    {data[KEY_VALUES[NODE].TITLE.id] &&
-                        <tspan key={Math.random()} x="50%"
-                               y={(data[KEY_VALUES[NODE].STROKE_WIDTH.id] || 1) + fontsize}
-                               dominantBaseline="middle" textAnchor="middle">{data.title}</tspan>
-                    }
-                    {data[KEY_VALUES[NODE].LABEL.id] &&
-                        data.label.split("\n").map((e, i) => {
-                            if (i !== 0) {
-                                return <tspan key={i} x="50%" dy={fontsize} dominantBaseline="middle"
-                                              textAnchor="middle">{e}</tspan>
-                            } else {
-                                return <tspan key={i} x="50%"
-                                              y={50 - ((data[KEY_VALUES[NODE].LABEL.id].split("\n").length - 1) * height / fontsize / 2) + "%"}
-                                              dominantBaseline="middle" textAnchor="middle">{e}</tspan>
-                            }
-                        })}
-
-                </text>
-            </svg>
+            {myCustomComponent}
 
             {
                 // TODO: niet hardcoded, deze waarden zijn al eens gedefinieerd in een hashmap
@@ -73,9 +101,9 @@ export default memo(({data, isConnectable}) => {
                      ["source", "top", "top-source"], ["source", "left", "left-source"],
                      ["target", "bottom", "bottom-target"], ["target", "right", "right-target"]*/
                 [["source", "right", "right"], ["source", "bottom", "bottom"],
-                    ["target", "left", "left"], ["target", "top", "top"],
-                    ["source", "top", "top"], ["source", "left", "left"],
-                    ["target", "bottom", "bottom"], ["target", "right", "right"]
+                ["target", "left", "left"], ["target", "top", "top"],
+                ["source", "top", "top"], ["source", "left", "left"],
+                ["target", "bottom", "bottom"], ["target", "right", "right"]
                 ].map((e, i) => {
                     return (
                         <Handle